shrink dns port to u16 as that should be enough

+++ pkg/agent/config.go
@@ -163,7 +163,7 @@ type Config struct {
 	EnableDNSTracking bool `env:"ENABLE_DNS_TRACKING" envDefault:"false"`
 	// DNSTrackingPort used to define which port the DNS service is mapped to at the pod level,
 	// so we can track DNS at the pod level
-	DNSTrackingPort int `env:"DNS_TRACKING_PORT" envDefault:"53"`
+	DNSTrackingPort uint16 `env:"DNS_TRACKING_PORT" envDefault:"53"`
 	// StaleEntriesEvictTimeout specifies the maximum duration that stale entries are kept
 	// before being deleted, default is 5 seconds.
 	StaleEntriesEvictTimeout time.Duration `env:"STALE_ENTRIES_EVICT_TIMEOUT" envDefault:"5s"`
